Refactor likelihood module structure
Rename gaussian_pointmass.py to _gaussian_pointmass.py and
update imports to use new private module naming convention:
- Update internal imports to use _gaussian and _base modules
- Export ConstGaussianPM and TrivialStatistic from public
  firecrown.likelihood API
- Update all example and test imports to use public API
- Update Makefile pre-commit target to include docs build
- Refactor test fixtures to use public reset() method
  instead of protected _reset() for TwoPointTheory

This change aligns with the module reorganization where
implementation modules use underscore prefix while
maintaining a clean public API through __init__.py exports.
