+++ examples/des_y1_3x2pt/factory_PM.py
@@ -4,11 +4,9 @@ import os
 
 import pyccl
 from firecrown.likelihood.factories import load_sacc_data
-from firecrown.likelihood.likelihood import NamedParameters
+from firecrown.likelihood import NamedParameters, TwoPoint, ConstGaussianPM
 import firecrown.likelihood.weak_lensing as wl
 import firecrown.likelihood.number_counts as nc
-from firecrown.likelihood.two_point import TwoPoint
-from firecrown.likelihood.gaussian_pointmass import ConstGaussianPM
 from firecrown.modeling_tools import ModelingTools
 from firecrown.ccl_factory import CCLFactory
