+++ firecrown/likelihood/_gaussian_pointmass.py
@@ -11,9 +11,8 @@ import numpy.typing as npt
 from scipy.integrate import simpson
 import pyccl
 
-from firecrown.likelihood.gaussian import ConstGaussian
-from firecrown.likelihood.statistic import Statistic
-
+from firecrown.likelihood._gaussian import ConstGaussian
+from firecrown.likelihood._base import Statistic
 
 # Default values for point mass marginalization
 DEFAULT_SIGMA_B = 10000.0
